Refactored StudentSubmission to extract helpers in to Utils and created an AuthenticatedUser class.
