+++ src/main/java/org/cs1331/gitsubmitter/Forker.java
@@ -1,9 +1,9 @@
 package org.cs1331.gitsubmitter;
 
-import java.util.Scanner;
 import java.io.File;
-import javax.swing.JPasswordField;
+import java.util.Scanner;
 import javax.swing.JOptionPane;
+import javax.swing.JPasswordField;
 
 public class Forker {
     private static Scanner scan = new Scanner(System.in);
@@ -16,28 +16,7 @@ public class Forker {
             System.out.println("java Forker -timedlab cs1331-timedlabs b1.txt");
             System.exit(0);
         }
-        String user = "";
-        String password = "";
-        String auth = "";
-        do {
-            if (!user.equals("")) {
-                System.err.println("Login failed, try again: ");
-            }
-            System.err.print("Username: ");
-            user = scan.nextLine();
-            System.err.println("Password Popup Displayed");
-            JPasswordField pf = new JPasswordField();
-            int res = JOptionPane.showConfirmDialog(null, pf, "Password:",
-                JOptionPane.OK_CANCEL_OPTION, JOptionPane.PLAIN_MESSAGE);
-            if (res == JOptionPane.OK_OPTION) {
-                password = new String(pf.getPassword());
-            } else {
-                System.exit(0);
-            }
-            //yes the colon
-            auth = java.util.Base64.getEncoder().encodeToString((
-                user + ":" + password).getBytes());
-        } while (!StudentSubmission.testAuth(auth));
+        AuthenticatedUser authenticatedUser = AuthenticatedUser.create();
 
         String repoSuffix = args[0];
         String org = args[1];
@@ -47,20 +26,23 @@ public class Forker {
             if (s2.equals("-d")) delete = true;
         }
         while (s.hasNext()) {
+            // WTH is u?
             String u = s.nextLine();
             if (delete) {
                 StudentSubmission ss =
-                    new StudentSubmission(user, auth, u + repoSuffix);
+                    new StudentSubmission(authenticatedUser, u + repoSuffix);
                 System.out.println(u);
                 System.out.println(ss.delete());
             } else {
-                StudentSubmission ss =
-                    new StudentSubmission(u, auth, u + repoSuffix);
-                System.out.println(u);
-                System.out.println(args.length > 2 ? ss.fork(org) : ss.fork());
-                new StudentSubmission(
-                    args.length > 2 ? org : user, auth, u + repoSuffix)
-                    .removeCollab(u);
+                System.out.println("TODO: Implement me. Not forked!");
+
+                // StudentSubmission ss =
+                //     new StudentSubmission(authenticatedUser, u + repoSuffix);
+                // System.out.println(u);
+                // System.out.println(args.length > 2 ? ss.fork(org) : ss.fork());
+                // new StudentSubmission(
+                //     args.length > 2 ? org : user, auth, u + repoSuffix)
+                //     .removeCollab(u);
             }
         }
     }
